backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2>=2.9
adrf
django-cors-headers
requests
rdflib

